% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{fmPlot}
\alias{fmPlot}
\title{fmPlot}
\usage{
fmPlot(dat, g, t)
}
\arguments{
\item{dat}{econFM output object}

\item{g}{character, group variable name}

\item{t}{integer}
}
\value{
plot
}
\description{
fmPlot
}
