% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{loessTM}
\alias{loessTM}
\title{Title}
\usage{
loessTM(
  dat,
  y,
  y_metric = 1000,
  log_base = 10,
  cpi_adjust = TRUE,
  time_var = "date",
  l = labs(title = "Pls add labels")
)
}
\arguments{
\item{dat}{tibble}

\item{y}{character, variable of interest name}

\item{y_metric, }{integer, multiplier (e.g. 1000)}

\item{log_base}{'natural' or 10}

\item{cpi_adjust}{logical}

\item{time_var}{character, time variable name}

\item{l}{label object (e.g. labs(title = 'hello'))}
}
\value{
ggplot
}
\description{
Title
}
