% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econFM.R
\name{econFM}
\alias{econFM}
\title{Title}
\usage{
econFM(
  dat,
  y,
  y_metric = 1000,
  log_base = 10,
  cpi_adjust = TRUE,
  time_var = "date",
  time_split = "none",
  time_split_cumulative = F,
  group_var = "disaster_subgroup",
  group = "",
  mc_method = "gft",
  mbin = 0.1
)
}
\arguments{
\item{dat}{data.frame tibble}

\item{y}{character}

\item{y_metric}{numeric}

\item{log_base}{10 or 'natural'}

\item{cpi_adjust}{logical}

\item{time_var}{character}

\item{time_split}{'median', 'half', 'none', or numeric quantile e.g. 0.2}

\item{time_split_cumulative}{logical}

\item{group_var}{character}

\item{group}{character vector}

\item{mc_method}{'gft', 'mode', or 'mbass'}

\item{mbin}{numeric}
}
\value{
list
}
\description{
Title
}
